@@ -1,109 +1,126 @@
 {
   "name": "webtorrent",
   "description": "Streaming torrent client",
-  "version": "1.8.18",
+  "version": "2.4.14",
   "author": {
     "name": "WebTorrent LLC",
     "email": "feross@webtorrent.io",
     "url": "https://webtorrent.io"
   },
+  "type": "module",
   "browser": {
-    "./lib/server.js": false,
     "./lib/conn-pool.js": false,
-    "./lib/utp.js": false,
-    "bittorrent-dht/client": false,
+    "./lib/utp.cjs": false,
+    "@silentbot1/nat-api": false,
+    "bittorrent-dht": false,
+    "crypto": false,
     "fs": false,
-    "fs-chunk-store": "memory-chunk-store",
+    "fs-chunk-store": "hybrid-chunk-store",
+    "http": false,
     "load-ip-set": false,
     "net": false,
     "os": false,
     "ut_pex": false
   },
-  "browserify": {
-    "transform": [
-      "package-json-versionify"
-    ]
-  },
   "bugs": {
     "url": "https://github.com/webtorrent/webtorrent/issues"
   },
   "chromeapp": {
-    "./lib/utp.js": false,
+    "./lib/utp.cjs": false,
+    "@silentbot1/nat-api": false,
+    "fs": false,
     "fs-chunk-store": "memory-chunk-store",
     "http": "@webtorrent/http-node",
+    "https": "@webtorrent/http-node",
     "load-ip-set": false,
     "net": "chrome-net",
     "os": false
   },
   "dependencies": {
+    "@silentbot1/nat-api": "^0.4.7",
+    "@thaunknown/simple-peer": "^10.0.10",
     "@webtorrent/http-node": "^1.3.0",
-    "addr-to-ip-port": "^1.5.4",
-    "bitfield": "^4.1.0",
-    "bittorrent-dht": "^10.0.2",
-    "bittorrent-protocol": "^3.5.5",
+    "addr-to-ip-port": "^2.0.0",
+    "bitfield": "^4.2.0",
+    "bittorrent-dht": "^11.0.7",
+    "bittorrent-protocol": "^4.1.14",
     "cache-chunk-store": "^3.2.2",
-    "chrome-net": "^3.3.4",
-    "chunk-store-stream": "^4.3.0",
+    "chunk-store-iterator": "^1.0.3",
     "cpus": "^1.0.3",
-    "create-torrent": "^5.0.2",
-    "debug": "^4.3.4",
-    "end-of-stream": "^1.4.4",
+    "create-torrent": "^6.0.18",
+    "cross-fetch-ponyfill": "^1.0.3",
+    "debug": "^4.3.6",
     "escape-html": "^1.0.3",
-    "fs-chunk-store": "^2.0.5",
+    "fs-chunk-store": "^4.1.0",
+    "hybrid-chunk-store": "^1.2.6",
     "immediate-chunk-store": "^2.2.0",
-    "load-ip-set": "^2.2.1",
-    "lt_donthave": "^1.0.1",
+    "join-async-iterator": "^1.1.1",
+    "load-ip-set": "^3.0.1",
+    "lt_donthave": "^2.0.2",
     "memory-chunk-store": "^1.3.5",
     "mime": "^3.0.0",
-    "multistream": "^4.1.0",
-    "package-json-versionify": "^1.0.4",
-    "parse-torrent": "^9.1.5",
+    "once": "^1.4.0",
+    "parse-torrent": "^11.0.17",
     "pump": "^3.0.0",
     "queue-microtask": "^1.2.3",
     "random-iterate": "^1.0.1",
-    "randombytes": "^2.1.0",
     "range-parser": "^1.2.1",
-    "render-media": "^4.1.0",
     "run-parallel": "^1.2.0",
     "run-parallel-limit": "^1.1.0",
-    "simple-concat": "^1.0.1",
-    "simple-get": "^4.0.1",
-    "simple-peer": "^9.11.1",
-    "simple-sha1": "^3.1.0",
     "speed-limiter": "^1.0.2",
-    "stream-to-blob": "^2.0.1",
-    "stream-to-blob-url": "^3.0.2",
-    "stream-with-known-length-to-buffer": "^1.0.4",
-    "throughput": "^1.0.0",
-    "torrent-discovery": "^9.4.10",
-    "torrent-piece": "^2.0.1",
+    "streamx": "2.17.0",
+    "throughput": "^1.0.1",
+    "torrent-discovery": "^11.0.8",
+    "torrent-piece": "^3.0.0",
+    "uint8-util": "^2.2.5",
     "unordered-array-remove": "^1.0.2",
-    "ut_metadata": "^3.5.2",
-    "ut_pex": "^3.0.2"
+    "ut_metadata": "^4.0.3",
+    "ut_pex": "^4.0.4"
   },
   "devDependencies": {
-    "@webtorrent/semantic-release-config": "1.0.7",
+    "@babel/core": "7.24.9",
+    "@babel/eslint-parser": "^7.25.1",
+    "@babel/plugin-syntax-import-assertions": "7.24.7",
+    "@babel/preset-env": "7.25.0",
+    "@webtorrent/semantic-release-config": "1.0.10",
     "airtap": "4.0.4",
     "airtap-manual": "1.0.0",
     "airtap-sauce": "1.1.2",
-    "babel-minify": "0.5.2",
-    "bittorrent-tracker": "9.18.6",
-    "browserify": "17.0.0",
+    "airtap-system": "^0.1.0",
+    "babelify": "10.0.0",
+    "bittorrent-tracker": "11.1.1",
+    "buffer": "^6.0.3",
+    "chrome-net": "^3.3.4",
+    "crypto-browserify": "^3.12.0",
     "disc": "1.3.3",
+    "eslint": "^8.57.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.2.0",
     "finalhandler": "1.2.0",
     "network-address": "1.1.2",
+    "pako": "^2.1.0",
+    "path-esm": "^1.0.0",
+    "querystring": "^0.2.1",
     "run-series": "1.1.9",
-    "semantic-release": "19.0.2",
+    "semantic-release": "22.0.12",
     "serve-static": "1.15.0",
-    "standard": "*",
-    "tape": "5.5.3",
+    "stream-browserify": "^3.0.0",
+    "tap-parser": "^16.0.1",
+    "tap-spec": "^5.0.0",
+    "tape": "^5.8.1",
+    "terser-webpack-plugin": "^5.3.10",
+    "timers-browserify": "^2.0.12",
+    "webpack": "^5.92.1",
+    "webpack-cli": "^5.1.4",
     "webtorrent-fixtures": "1.7.5"
   },
   "optionalDependencies": {
     "utp-native": "^2.5.3"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "funding": [
     {
