@@ -1,27 +1,23 @@
-const EventEmitter = require('events')
-const { PassThrough } = require('stream')
-const path = require('path')
-const render = require('render-media')
-const streamToBlob = require('stream-to-blob')
-const streamToBlobURL = require('stream-to-blob-url')
-const streamToBuffer = require('stream-with-known-length-to-buffer')
-const queueMicrotask = require('queue-microtask')
-const rangeParser = require('range-parser')
-const mime = require('mime')
-const eos = require('end-of-stream')
-const FileStream = require('./file-stream.js')
+import EventEmitter from 'events'
+import { Readable } from 'streamx'
+import { chunkStoreRead } from 'chunk-store-iterator'
+import mime from 'mime/lite.js'
+import FileIterator from './file-iterator.js'
 
-class File extends EventEmitter {
+export default class File extends EventEmitter {
   constructor (torrent, file) {
     super()
 
     this._torrent = torrent
     this._destroyed = false
     this._fileStreams = new Set()
+    this._iterators = new Set()
 
     this.name = file.name
     this.path = file.path
     this.length = file.length
+    this.size = file.length
+    this.type = mime.getType(this.name) || 'application/octet-stream'
     this.offset = file.offset
 
     this.done = false
@@ -37,7 +33,7 @@ class File extends EventEmitter {
       this.emit('done')
     }
 
-    this._serviceWorker = torrent.client.serviceWorker
+    this._client = torrent.client
   }
 
   get downloaded () {
@@ -95,19 +91,33 @@ class File extends EventEmitter {
 
   deselect () {
     if (this.length === 0) return
-    this._torrent.deselect(this._startPiece, this._endPiece, false)
+    this._torrent.deselect(this._startPiece, this._endPiece)
+  }
+
+  [Symbol.asyncIterator] (opts = {}) {
+    if (this.length === 0) return (async function * empty () {})()
+
+    const { start = 0 } = opts ?? {}
+    const end = (opts?.end && opts.end < this.length)
+      ? opts.end
+      : this.length - 1
+
+    if (this.done) {
+      return chunkStoreRead(this._torrent.store, { offset: start + this.offset, length: end - start + 1 })
+    }
+
+    const iterator = new FileIterator(this, { start, end })
+    this._iterators.add(iterator)
+    iterator.once('return', () => {
+      this._iterators.delete(iterator)
+    })
+
+    return iterator
   }
 
   createReadStream (opts) {
-    if (this.length === 0) {
-      const empty = new PassThrough()
-      queueMicrotask(() => {
-        empty.end()
-      })
-      return empty
-    }
-
-    const fileStream = new FileStream(this, opts)
+    const iterator = this[Symbol.asyncIterator](opts)
+    const fileStream = Readable.from(iterator)
 
     this._fileStreams.add(fileStream)
     fileStream.once('close', () => {
@@ -117,111 +127,52 @@ class File extends EventEmitter {
     return fileStream
   }
 
-  getBuffer (cb) {
-    streamToBuffer(this.createReadStream(), this.length, cb)
+  async arrayBuffer (opts) {
+    const data = new Uint8Array(this.length)
+    let offset = 0
+    for await (const chunk of this[Symbol.asyncIterator](opts)) {
+      data.set(chunk, offset)
+      offset += chunk.length
+    }
+    return data.buffer
   }
 
-  getBlob (cb) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    streamToBlob(this.createReadStream(), this._getMimeType())
-      .then(
-        blob => cb(null, blob),
-        err => cb(err)
-      )
+  async blob (opts) {
+    return new Blob([await this.arrayBuffer(opts)], { type: this.type })
   }
 
-  getBlobURL (cb) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    streamToBlobURL(this.createReadStream(), this._getMimeType())
-      .then(
-        blobUrl => cb(null, blobUrl),
-        err => cb(err)
-      )
-  }
-
-  appendTo (elem, opts, cb) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    render.append(this, elem, opts, cb)
-  }
-
-  renderTo (elem, opts, cb) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    render.render(this, elem, opts, cb)
-  }
-
-  _serve (req) {
-    const res = {
-      status: 200,
-      headers: {
-        // Support range-requests
-        'Accept-Ranges': 'bytes',
-        'Content-Type': mime.getType(this.name),
-        'Cache-Control': 'no-cache, no-store, must-revalidate, max-age=0',
-        Expires: '0'
+  stream (opts) {
+    let iterator
+    return new ReadableStream({
+      start: () => {
+        iterator = this[Symbol.asyncIterator](opts)
       },
-      body: req.method === 'HEAD' ? '' : 'STREAM'
-    }
-    // force the browser to download the file if if it's opened in a new tab
-    if (req.destination === 'document') {
-      res.headers['Content-Type'] = 'application/octet-stream'
-      res.headers['Content-Disposition'] = 'attachment'
-      res.body = 'DOWNLOAD'
-    }
-
-    // `rangeParser` returns an array of ranges, or an error code (number) if
-    // there was an error parsing the range.
-    let range = rangeParser(this.length, req.headers.range || '')
-
-    if (range.constructor === Array) {
-      res.status = 206 // indicates that range-request was understood
-
-      // no support for multi-range request, just use the first range
-      range = range[0]
-
-      res.headers['Content-Range'] = `bytes ${range.start}-${range.end}/${this.length}`
-      res.headers['Content-Length'] = `${range.end - range.start + 1}`
-    } else {
-      res.headers['Content-Length'] = this.length
-    }
-
-    const stream = req.method === 'GET' && this.createReadStream(range)
-
-    let pipe = null
-    if (stream) {
-      this.emit('stream', { stream, req, file: this }, piped => {
-        pipe = piped
-
-        // piped stream might not close the original filestream on close/error, this is agressive but necessary
-        eos(piped, () => {
-          if (piped) piped.destroy()
-          stream.destroy()
-        })
-      })
-    }
-
-    return [res, pipe || stream, pipe && stream]
+      async pull (controller) {
+        const { value, done } = await iterator.next()
+        if (done) {
+          controller.close()
+        } else {
+          controller.enqueue(value)
+        }
+      },
+      cancel () {
+        iterator.return()
+      }
+    })
   }
 
-  getStreamURL (cb = () => {}) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    if (!this._serviceWorker) throw new Error('No worker registered')
-    if (this._serviceWorker.state !== 'activated') throw new Error('Worker isn\'t activated')
-    const workerPath = this._serviceWorker.scriptURL.slice(0, this._serviceWorker.scriptURL.lastIndexOf('/') + 1).slice(window.location.origin.length)
-    const url = `${workerPath}webtorrent/${this._torrent.infoHash}/${encodeURI(this.path)}`
-    cb(null, url)
+  get streamURL () {
+    if (!this._client._server) throw new Error('No server created')
+    return `${this._client._server.pathname}/${this._torrent.infoHash}/${this.path}`
   }
 
-  streamTo (elem, cb = () => {}) {
-    if (typeof window === 'undefined') throw new Error('browser-only method')
-    if (!this._serviceWorker) throw new Error('No worker registered')
-    if (this._serviceWorker.state !== 'activated') throw new Error('Worker isn\'t activated')
-    const workerPath = this._serviceWorker.scriptURL.slice(0, this._serviceWorker.scriptURL.lastIndexOf('/') + 1).slice(window.location.origin.length)
-    elem.src = `${workerPath}webtorrent/${this._torrent.infoHash}/${encodeURI(this.path)}`
-    cb(null, elem)
+  streamTo (elem) {
+    elem.src = this.streamURL
+    return elem
   }
 
-  _getMimeType () {
-    return render.mime[path.extname(this.name).toLowerCase()]
+  includes (piece) {
+    return this._startPiece <= piece && this._endPiece >= piece
   }
 
   _destroy () {
@@ -232,7 +183,9 @@ class File extends EventEmitter {
       fileStream.destroy()
     }
     this._fileStreams.clear()
+    for (const iterator of this._iterators) {
+      iterator.destroy()
+    }
+    this._iterators.clear()
   }
 }
-
-module.exports = File
