@@ -1,122 +1,35 @@
-const EventEmitter = require('events')
-const { Transform } = require('stream')
-const arrayRemove = require('unordered-array-remove')
-const debugFactory = require('debug')
-const Wire = require('bittorrent-protocol')
+import EventEmitter from 'events'
+import { Transform, pipeline } from 'streamx'
+import arrayRemove from 'unordered-array-remove'
+import debugFactory from 'debug'
+import Wire from 'bittorrent-protocol'
 
-const CONNECT_TIMEOUT_TCP = 5000
-const CONNECT_TIMEOUT_UTP = 5000
-const CONNECT_TIMEOUT_WEBRTC = 25000
-const HANDSHAKE_TIMEOUT = 25000
+const CONNECT_TIMEOUT_TCP = 5_000
+const CONNECT_TIMEOUT_UTP = 5_000
+const CONNECT_TIMEOUT_WEBRTC = 25_000
+const HANDSHAKE_TIMEOUT = 25_000
+
+// Types of peers
+const TYPE_TCP_INCOMING = 'tcpIncoming'
+const TYPE_TCP_OUTGOING = 'tcpOutgoing'
+const TYPE_UTP_INCOMING = 'utpIncoming'
+const TYPE_UTP_OUTGOING = 'utpOutgoing'
+const TYPE_WEBRTC = 'webrtc'
+const TYPE_WEBSEED = 'webSeed'
+
+// Source used to obtain the peer
+const SOURCE_MANUAL = 'manual'
+const SOURCE_TRACKER = 'tracker'
+const SOURCE_DHT = 'dht'
+const SOURCE_LSD = 'lsd'
+const SOURCE_UT_PEX = 'ut_pex'
+
 const debug = debugFactory('webtorrent:peer')
 
 let secure = false
 
-exports.enableSecure = () => {
+export const enableSecure = () => {
   secure = true
-}
-
-/**
- * WebRTC peer connections start out connected, because WebRTC peers require an
- * "introduction" (i.e. WebRTC signaling), and there's no equivalent to an IP address
- * that lets you refer to a WebRTC endpoint.
- */
-exports.createWebRTCPeer = (conn, swarm, throttleGroups) => {
-  const peer = new Peer(conn.id, 'webrtc')
-  peer.conn = conn
-  peer.swarm = swarm
-  peer.throttleGroups = throttleGroups
-
-  if (peer.conn.connected) {
-    peer.onConnect()
-  } else {
-    const cleanup = () => {
-      peer.conn.removeListener('connect', onConnect)
-      peer.conn.removeListener('error', onError)
-    }
-    const onConnect = () => {
-      cleanup()
-      peer.onConnect()
-    }
-    const onError = err => {
-      cleanup()
-      peer.destroy(err)
-    }
-    peer.conn.once('connect', onConnect)
-    peer.conn.once('error', onError)
-    peer.startConnectTimeout()
-  }
-
-  return peer
-}
-
-/**
- * Incoming TCP peers start out connected, because the remote peer connected to the
- * listening port of the TCP server. Until the remote peer sends a handshake, we don't
- * know what swarm the connection is intended for.
- */
-exports.createTCPIncomingPeer = (conn, throttleGroups) => {
-  return _createIncomingPeer(conn, 'tcpIncoming', throttleGroups)
-}
-
-/**
- * Incoming uTP peers start out connected, because the remote peer connected to the
- * listening port of the uTP server. Until the remote peer sends a handshake, we don't
- * know what swarm the connection is intended for.
- */
-exports.createUTPIncomingPeer = (conn, throttleGroups) => {
-  return _createIncomingPeer(conn, 'utpIncoming', throttleGroups)
-}
-
-/**
- * Outgoing TCP peers start out with just an IP address. At some point (when there is an
- * available connection), the client can attempt to connect to the address.
- */
-exports.createTCPOutgoingPeer = (addr, swarm, throttleGroups) => {
-  return _createOutgoingPeer(addr, swarm, 'tcpOutgoing', throttleGroups)
-}
-
-/**
- * Outgoing uTP peers start out with just an IP address. At some point (when there is an
- * available connection), the client can attempt to connect to the address.
- */
-exports.createUTPOutgoingPeer = (addr, swarm, throttleGroups) => {
-  return _createOutgoingPeer(addr, swarm, 'utpOutgoing', throttleGroups)
-}
-
-const _createIncomingPeer = (conn, type, throttleGroups) => {
-  const addr = `${conn.remoteAddress}:${conn.remotePort}`
-  const peer = new Peer(addr, type)
-  peer.conn = conn
-  peer.addr = addr
-  peer.throttleGroups = throttleGroups
-
-  peer.onConnect()
-
-  return peer
-}
-
-const _createOutgoingPeer = (addr, swarm, type, throttleGroups) => {
-  const peer = new Peer(addr, type)
-  peer.addr = addr
-  peer.swarm = swarm
-  peer.throttleGroups = throttleGroups
-
-  return peer
-}
-
-/**
- * Peer that represents a Web Seed (BEP17 / BEP19).
- */
-exports.createWebSeedPeer = (conn, id, swarm, throttleGroups) => {
-  const peer = new Peer(id, 'webSeed')
-  peer.swarm = swarm
-  peer.conn = conn
-  peer.throttleGroups = throttleGroups
-
-  peer.onConnect()
-
-  return peer
 }
 
 /**
@@ -125,7 +38,7 @@ exports.createWebSeedPeer = (conn, id, swarm, throttleGroups) => {
  * @param {string} id "ip:port" string, peer id (for WebRTC peers), or url (for Web Seeds)
  * @param {string} type the type of the peer
  */
-class Peer extends EventEmitter {
+export default class Peer extends EventEmitter {
   constructor (id, type) {
     super()
 
@@ -138,6 +51,7 @@ class Peer extends EventEmitter {
     this.conn = null
     this.swarm = null
     this.wire = null
+    this.source = null
 
     this.connected = false
     this.destroyed = false
@@ -267,25 +181,27 @@ class Peer extends EventEmitter {
 
   setThrottlePipes () {
     const self = this
-    this.conn
-      .pipe(this.throttleGroups.down.throttle())
-      .pipe(new Transform({
-        transform (chunk, _, callback) {
+    pipeline(
+      this.conn,
+      this.throttleGroups.down.throttle(),
+      new Transform({
+        transform (chunk, callback) {
           self.emit('download', chunk.length)
           if (self.destroyed) return
           callback(null, chunk)
         }
-      }))
-      .pipe(this.wire)
-      .pipe(this.throttleGroups.up.throttle())
-      .pipe(new Transform({
-        transform (chunk, _, callback) {
+      }),
+      this.wire,
+      this.throttleGroups.up.throttle(),
+      new Transform({
+        transform (chunk, callback) {
           self.emit('upload', chunk.length)
           if (self.destroyed) return
           callback(null, chunk)
         }
-      }))
-      .pipe(this.conn)
+      }),
+      this.conn
+    )
   }
 
   /**
@@ -385,4 +301,122 @@ class Peer extends EventEmitter {
     if (wire) wire.destroy()
     if (swarm) swarm.removePeer(this.id)
   }
+}
+
+Peer.TYPE_TCP_INCOMING = TYPE_TCP_INCOMING
+Peer.TYPE_TCP_OUTGOING = TYPE_TCP_OUTGOING
+Peer.TYPE_UTP_INCOMING = TYPE_UTP_INCOMING
+Peer.TYPE_UTP_OUTGOING = TYPE_UTP_OUTGOING
+Peer.TYPE_WEBRTC = TYPE_WEBRTC
+Peer.TYPE_WEBSEED = TYPE_WEBSEED
+
+Peer.SOURCE_MANUAL = SOURCE_MANUAL
+Peer.SOURCE_TRACKER = SOURCE_TRACKER
+Peer.SOURCE_DHT = SOURCE_DHT
+Peer.SOURCE_LSD = SOURCE_LSD
+Peer.SOURCE_UT_PEX = SOURCE_UT_PEX
+
+/**
+ * WebRTC peer connections start out connected, because WebRTC peers require an
+ * "introduction" (i.e. WebRTC signaling), and there's no equivalent to an IP address
+ * that lets you refer to a WebRTC endpoint.
+ */
+Peer.createWebRTCPeer = (conn, swarm, throttleGroups) => {
+  const peer = new Peer(conn.id, 'webrtc')
+  peer.conn = conn
+  peer.swarm = swarm
+  peer.throttleGroups = throttleGroups
+
+  if (peer.conn.connected) {
+    peer.onConnect()
+  } else {
+    const cleanup = () => {
+      peer.conn.removeListener('connect', onConnect)
+      peer.conn.removeListener('error', onError)
+    }
+    const onConnect = () => {
+      cleanup()
+      peer.onConnect()
+    }
+    const onError = err => {
+      cleanup()
+      peer.destroy(err)
+    }
+    peer.conn.once('connect', onConnect)
+    peer.conn.once('error', onError)
+    peer.startConnectTimeout()
+  }
+
+  return peer
+}
+
+/**
+ * Incoming TCP peers start out connected, because the remote peer connected to the
+ * listening port of the TCP server. Until the remote peer sends a handshake, we don't
+ * know what swarm the connection is intended for.
+ */
+Peer.createTCPIncomingPeer = (conn, throttleGroups) => {
+  return Peer._createIncomingPeer(conn, TYPE_TCP_INCOMING, throttleGroups)
+}
+
+/**
+ * Incoming uTP peers start out connected, because the remote peer connected to the
+ * listening port of the uTP server. Until the remote peer sends a handshake, we don't
+ * know what swarm the connection is intended for.
+ */
+Peer.createUTPIncomingPeer = (conn, throttleGroups) => {
+  return Peer._createIncomingPeer(conn, TYPE_UTP_INCOMING, throttleGroups)
+}
+
+/**
+ * Outgoing TCP peers start out with just an IP address. At some point (when there is an
+ * available connection), the client can attempt to connect to the address.
+ */
+Peer.createTCPOutgoingPeer = (addr, swarm, throttleGroups) => {
+  return Peer._createOutgoingPeer(addr, swarm, TYPE_TCP_OUTGOING, throttleGroups)
+}
+
+/**
+ * Outgoing uTP peers start out with just an IP address. At some point (when there is an
+ * available connection), the client can attempt to connect to the address.
+ */
+Peer.createUTPOutgoingPeer = (addr, swarm, throttleGroups) => {
+  return Peer._createOutgoingPeer(addr, swarm, TYPE_UTP_OUTGOING, throttleGroups)
+}
+
+Peer._createIncomingPeer = (conn, type, throttleGroups) => {
+  const addr = `${conn.remoteAddress}:${conn.remotePort}`
+  const peer = new Peer(addr, type)
+  peer.conn = conn
+  peer.addr = addr
+  peer.throttleGroups = throttleGroups
+
+  peer.onConnect()
+
+  return peer
+}
+
+Peer._createOutgoingPeer = (addr, swarm, type, throttleGroups) => {
+  const peer = new Peer(addr, type)
+  peer.addr = addr
+  peer.swarm = swarm
+  peer.throttleGroups = throttleGroups
+
+  return peer
+}
+
+/**
+ * Peer that represents a Web Seed (BEP17 / BEP19).
+ */
+
+Peer.createWebSeedPeer = (conn, id, swarm, throttleGroups) => {
+  const peer = new Peer(id, TYPE_WEBSEED)
+
+  peer.swarm = swarm
+  peer.conn = conn
+  peer.throttleGroups = throttleGroups
+
+  peer.onConnect()
+
+  return peer
 }
