@@ -1,249 +1,424 @@
-const http = require('http')
-const escapeHtml = require('escape-html')
-const mime = require('mime')
-const pump = require('pump')
-const rangeParser = require('range-parser')
-const queueMicrotask = require('queue-microtask')
+import http from 'http'
+import escapeHtml from 'escape-html'
+import pump from 'pump'
+import rangeParser from 'range-parser'
+import queueMicrotask from 'queue-microtask'
+import { Readable } from 'streamx'
 
-function Server (torrent, opts = {}) {
-  const server = http.createServer()
-  if (!opts.origin) opts.origin = '*' // allow all origins by default
+const keepAliveTime = 20000
 
-  const sockets = new Set()
-  const pendingReady = new Set()
-  let closed = false
-  const _listen = server.listen
-  const _close = server.close
-
-  server.listen = (...args) => {
-    closed = false
-    server.on('connection', onConnection)
-    server.on('request', onRequest)
-    return _listen.apply(server, args)
+class ServerBase {
+  constructor (client, opts = {}) {
+    this.client = client
+    if (!opts.origin) opts.origin = '*' // allow all origins by default
+    this.opts = opts
+    this.pendingReady = new Set()
   }
 
-  server.close = cb => {
-    closed = true
-    server.removeListener('connection', onConnection)
-    server.removeListener('request', onRequest)
-    pendingReady.forEach(onReady => {
-      torrent.removeListener('ready', onReady)
-    })
-    pendingReady.clear()
-    _close.call(server, cb)
+  static serveIndexPage (res, torrents, pathname) {
+    const listHtml = torrents
+      .map(torrent => (
+      `<li>
+        <a href="${escapeHtml(pathname)}/${torrent.infoHash}">
+          ${escapeHtml(torrent.name)}
+        </a>
+        (${escapeHtml(torrent.length)} bytes)
+      </li>`
+      ))
+      .join('<br>')
+
+    res.status = 200
+    res.headers['Content-Type'] = 'text/html'
+    res.body = getPageHTML(
+      'WebTorrent',
+      `<h1>WebTorrent</h1>
+       <ol>${listHtml}</ol>`
+    )
+
+    return res
   }
 
-  server.destroy = cb => {
-    sockets.forEach(socket => {
-      socket.destroy()
-    })
-
-    // Only call `server.close` if user has not called it already
-    if (!cb) cb = () => {}
-    if (closed) queueMicrotask(cb)
-    else server.close(cb)
-    torrent = null
-  }
-
-  function isOriginAllowed (req) {
+  isOriginAllowed (req) {
     // When `origin` option is `false`, deny all cross-origin requests
-    if (opts.origin === false) return false
-
-    // Requests without an 'Origin' header are not actually cross-origin, so just
-    // deny them
-    if (req.headers.origin == null) return false
+    if (this.opts.origin === false) return false
 
     // The user allowed all origins
-    if (opts.origin === '*') return true
+    if (this.opts.origin === '*') return true
 
     // Allow requests where the 'Origin' header matches the `opts.origin` setting
-    return req.headers.origin === opts.origin
+    return req.headers.origin === this.opts.origin
   }
 
-  function onConnection (socket) {
-    socket.setTimeout(36000000)
-    sockets.add(socket)
-    socket.once('close', () => {
-      sockets.delete(socket)
-    })
+  static serveMethodNotAllowed (res) {
+    res.status = 405
+    res.headers['Content-Type'] = 'text/html'
+
+    res.body = getPageHTML(
+      '405 - Method Not Allowed',
+      '<h1>405 - Method Not Allowed</h1>'
+    )
+
+    return res
   }
 
-  function onRequest (req, res) {
-    // If a 'hostname' string is specified, deny requests with a 'Host'
-    // header that does not match the origin of the torrent server to prevent
-    // DNS rebinding attacks.
-    if (opts.hostname && req.headers.host !== `${opts.hostname}:${server.address().port}`) {
-      return req.destroy()
+  static serve404Page (res) {
+    res.status = 404
+    res.headers['Content-Type'] = 'text/html'
+
+    res.body = getPageHTML(
+      '404 - Not Found',
+      '<h1>404 - Not Found</h1>'
+    )
+    return res
+  }
+
+  static serveTorrentPage (torrent, res, pathname) {
+    const listHtml = torrent.files
+      .map(file => (
+      `<li>
+        <a href="${escapeHtml(pathname)}/${torrent.infoHash}/${escapeHtml(file.path)}">
+          ${escapeHtml(file.path)}
+        </a>
+        (${escapeHtml(file.length)} bytes)
+      </li>`
+      ))
+      .join('<br>')
+
+    res.status = 200
+    res.headers['Content-Type'] = 'text/html'
+
+    res.body = getPageHTML(
+      `${escapeHtml(torrent.name)} - WebTorrent`,
+      `<h1>${escapeHtml(torrent.name)}</h1>
+      <ol>${listHtml}</ol>`
+    )
+
+    return res
+  }
+
+  static serveOptionsRequest (req, res) {
+    res.status = 204 // no content
+    res.headers['Access-Control-Max-Age'] = '600'
+    res.headers['Access-Control-Allow-Methods'] = 'GET,HEAD'
+
+    if (req.headers['access-control-request-headers']) {
+      res.headers['Access-Control-Allow-Headers'] = req.headers['access-control-request-headers']
+    }
+    return res
+  }
+
+  static serveFile (file, req, res) {
+    res.status = 200
+
+    // Disable caching as data is local anyways
+    res.headers.Expires = '0'
+    res.headers['Cache-Control'] = 'no-cache, no-store, must-revalidate, max-age=0'
+    // Support range-requests
+    res.headers['Accept-Ranges'] = 'bytes'
+    res.headers['Content-Type'] = file.type
+    // Support DLNA streaming
+    res.headers['transferMode.dlna.org'] = 'Streaming'
+    res.headers['contentFeatures.dlna.org'] = 'DLNA.ORG_OP=01;DLNA.ORG_CI=0;DLNA.ORG_FLAGS=01700000000000000000000000000000'
+
+    // Force the browser to download the file if if it's opened in a new tab
+    // Set name of file (for "Save Page As..." dialog)
+    if (req.destination === 'document') {
+      res.headers['Content-Type'] = 'application/octet-stream'
+      res.headers['Content-Disposition'] = `attachment; filename*=UTF-8''${encodeRFC5987(file.name)}`
+      res.body = 'DOWNLOAD'
+    } else {
+      res.headers['Content-Disposition'] = `inline; filename*=UTF-8''${encodeRFC5987(file.name)}`
     }
 
-    const pathname = new URL(req.url, 'http://example.com').pathname
+    // `rangeParser` returns an array of ranges, or an error code (number) if
+    // there was an error parsing the range.
+    let range = rangeParser(file.length, req.headers.range || '')
+
+    if (Array.isArray(range)) {
+      res.status = 206 // indicates that range-request was understood
+
+      // no support for multi-range request, just use the first range
+      range = range[0]
+
+      res.headers['Content-Range'] = `bytes ${range.start}-${range.end}/${file.length}`
+
+      res.headers['Content-Length'] = range.end - range.start + 1
+    } else {
+      res.statusCode = 200
+      range = null
+      res.headers['Content-Length'] = file.length
+    }
+
+    if (req.method === 'GET') {
+      const iterator = file[Symbol.asyncIterator](range)
+      let transform = null
+      file.emit('iterator', { iterator, req, file }, target => {
+        transform = target
+      })
+
+      const stream = Readable.from(transform || iterator)
+      let pipe = null
+      file.emit('stream', { stream, req, file }, target => {
+        pipe = pump(stream, target)
+      })
+
+      res.body = pipe || stream
+    } else {
+      res.body = false
+    }
+    return res
+  }
+
+  async onRequest (req, cb) {
+    let pathname = new URL(req.url, 'http://example.com').pathname
+    pathname = pathname.slice(pathname.indexOf(this.pathname) + this.pathname.length + 1)
+
+    const res = {
+      headers: {
+        // Prevent browser mime-type sniffing
+        'X-Content-Type-Options': 'nosniff',
+        // Defense-in-depth: Set a strict Content Security Policy to mitigate XSS
+        'Content-Security-Policy': "base-uri 'none'; frame-ancestors 'none'; form-action 'none';"
+      }
+    }
 
     // Allow cross-origin requests (CORS)
-    if (isOriginAllowed(req)) {
-      res.setHeader('Access-Control-Allow-Origin', req.headers.origin)
+    if (this.isOriginAllowed(req)) {
+      res.headers['Access-Control-Allow-Origin'] = this.opts.origin === '*' ? '*' : req.headers.origin
     }
 
-    // Prevent browser mime-type sniffing
-    res.setHeader('X-Content-Type-Options', 'nosniff')
-
-    // Defense-in-depth: Set a strict Content Security Policy to mitigate XSS
-    res.setHeader('Content-Security-Policy', "base-uri 'none'; default-src 'none'; frame-ancestors 'none'; form-action 'none';")
-
-    if (pathname === '/favicon.ico') {
-      return serve404Page()
+    if (pathname === 'favicon.ico') {
+      return cb(ServerBase.serve404Page(res))
     }
 
     // Allow CORS requests to specify arbitrary headers, e.g. 'Range',
     // by responding to the OPTIONS preflight request with the specified
     // origin and requested headers.
     if (req.method === 'OPTIONS') {
-      if (isOriginAllowed(req)) return serveOptionsRequest()
-      else return serveMethodNotAllowed()
+      if (this.isOriginAllowed(req)) return cb(ServerBase.serveOptionsRequest(req, res))
+      else return cb(ServerBase.serveMethodNotAllowed(res))
+    }
+
+    const onReady = async () => {
+      this.pendingReady.delete(onReady)
+      const res = await handleRequest()
+      cb(res)
+    }
+
+    const handleRequest = async () => {
+      if (pathname === '') {
+        return ServerBase.serveIndexPage(res, this.client.torrents, this.pathname)
+      }
+
+      let [infoHash, ...filePath] = pathname.split('/')
+      filePath = decodeURI(filePath.join('/'))
+
+      const torrent = await this.client.get(infoHash)
+      if (!infoHash || !torrent) {
+        return ServerBase.serve404Page(res)
+      }
+
+      if (!filePath) {
+        return ServerBase.serveTorrentPage(torrent, res, this.pathname)
+      }
+
+      const file = torrent.files.find(file => file.path.replace(/\\/g, '/') === filePath)
+      if (!file) {
+        return ServerBase.serve404Page(res)
+      }
+      return ServerBase.serveFile(file, req, res)
     }
 
     if (req.method === 'GET' || req.method === 'HEAD') {
-      if (torrent.ready) {
-        return handleRequest()
+      if (this.client.ready) {
+        const res = await handleRequest()
+        return cb(res)
       } else {
-        pendingReady.add(onReady)
-        torrent.once('ready', onReady)
+        this.pendingReady.add(onReady)
+        this.client.once('ready', onReady)
         return
       }
     }
 
-    return serveMethodNotAllowed()
-
-    function serveOptionsRequest () {
-      res.statusCode = 204 // no content
-      res.setHeader('Access-Control-Max-Age', '600')
-      res.setHeader('Access-Control-Allow-Methods', 'GET,HEAD')
-
-      if (req.headers['access-control-request-headers']) {
-        res.setHeader(
-          'Access-Control-Allow-Headers',
-          req.headers['access-control-request-headers']
-        )
-      }
-      res.end()
-    }
-
-    function onReady () {
-      pendingReady.delete(onReady)
-      handleRequest()
-    }
-
-    function handleRequest () {
-      if (pathname === '/') {
-        return serveIndexPage()
-      }
-
-      const index = Number(pathname.split('/')[1])
-      if (Number.isNaN(index) || index >= torrent.files.length) {
-        return serve404Page()
-      }
-
-      const file = torrent.files[index]
-      serveFile(file)
-    }
-
-    function serveIndexPage () {
-      res.statusCode = 200
-      res.setHeader('Content-Type', 'text/html')
-
-      const listHtml = torrent.files
-        .map((file, i) => (
-          `<li>
-            <a
-              download="${escapeHtml(file.name)}"
-              href="${escapeHtml(i)}/${escapeHtml(file.name)}"
-            >
-              ${escapeHtml(file.path)}
-            </a>
-            (${escapeHtml(file.length)} bytes)
-          </li>`
-        ))
-        .join('<br>')
-
-      const html = getPageHTML(
-        `${escapeHtml(torrent.name)} - WebTorrent`,
-        `
-          <h1>${escapeHtml(torrent.name)}</h1>
-          <ol>${listHtml}</ol>
-        `
-      )
-      res.end(html)
-    }
-
-    function serve404Page () {
-      res.statusCode = 404
-      res.setHeader('Content-Type', 'text/html')
-
-      const html = getPageHTML(
-        '404 - Not Found',
-        '<h1>404 - Not Found</h1>'
-      )
-      res.end(html)
-    }
-
-    function serveFile (file) {
-      res.setHeader('Content-Type', mime.getType(file.name) || 'application/octet-stream')
-
-      // Support range-requests
-      res.setHeader('Accept-Ranges', 'bytes')
-
-      // Set name of file (for "Save Page As..." dialog)
-      res.setHeader(
-        'Content-Disposition',
-        `inline; filename*=UTF-8''${encodeRFC5987(file.name)}`
-      )
-
-      // Support DLNA streaming
-      res.setHeader('transferMode.dlna.org', 'Streaming')
-      res.setHeader(
-        'contentFeatures.dlna.org',
-        'DLNA.ORG_OP=01;DLNA.ORG_CI=0;DLNA.ORG_FLAGS=01700000000000000000000000000000'
-      )
-
-      // `rangeParser` returns an array of ranges, or an error code (number) if
-      // there was an error parsing the range.
-      let range = rangeParser(file.length, req.headers.range || '')
-
-      if (Array.isArray(range)) {
-        res.statusCode = 206 // indicates that range-request was understood
-
-        // no support for multi-range request, just use the first range
-        range = range[0]
-
-        res.setHeader(
-          'Content-Range',
-          `bytes ${range.start}-${range.end}/${file.length}`
-        )
-        res.setHeader('Content-Length', range.end - range.start + 1)
-      } else {
-        res.statusCode = 200
-        range = null
-        res.setHeader('Content-Length', file.length)
-      }
-
-      if (req.method === 'HEAD') {
-        return res.end()
-      }
-
-      pump(file.createReadStream(range), res)
-    }
-
-    function serveMethodNotAllowed () {
-      res.statusCode = 405
-      res.setHeader('Content-Type', 'text/html')
-      const html = getPageHTML(
-        '405 - Method Not Allowed',
-        '<h1>405 - Method Not Allowed</h1>'
-      )
-      res.end(html)
-    }
+    return cb(ServerBase.serveMethodNotAllowed(res))
   }
 
-  return server
+  close (cb = () => {}) {
+    this.closed = true
+    this.pendingReady.forEach(onReady => {
+      this.client.removeListener('ready', onReady)
+    })
+    this.pendingReady.clear()
+    queueMicrotask(cb)
+  }
+
+  destroy (cb = () => {}) {
+    // Only call `server.close` if user has not called it already
+    if (this.closed) queueMicrotask(cb)
+    else this.close(cb)
+    this.client = null
+  }
+}
+
+class NodeServer extends ServerBase {
+  constructor (client, opts) {
+    super(client, opts)
+
+    this.server = http.createServer()
+    this._listen = this.server.listen
+    this.server.listen = this.listen.bind(this)
+    this._close = this.server.close
+    this.server.close = this.close.bind(this)
+
+    this.sockets = new Set()
+    this.closed = false
+    this.pathname = opts?.pathname || '/webtorrent'
+  }
+
+  wrapRequest (req, res) {
+    // If a 'hostname' string is specified, deny requests with a 'Host'
+    // header that does not match the origin of the torrent server to prevent
+    // DNS rebinding attacks.
+    if (this.opts.hostname && req.headers.host !== `${this.opts.hostname}:${this.server.address().port}`) {
+      return req.destroy()
+    }
+
+    if (!new URL(req.url, 'http://example.com').pathname.startsWith(this.pathname)) {
+      return req.destroy()
+    }
+
+    this.onRequest(req, ({ status, headers, body }) => {
+      res.writeHead(status, headers)
+
+      if (!!body._readableState || !!body._writableState) { // this is probably a bad way of checking? idk
+        pump(body, res)
+      } else {
+        res.end(body)
+      }
+    })
+  }
+
+  onConnection (socket) {
+    socket.setTimeout(36000000)
+    this.sockets.add(socket)
+    socket.once('close', () => {
+      this.sockets.delete(socket)
+    })
+  }
+
+  address () {
+    return this.server.address()
+  }
+
+  listen (...args) {
+    this.closed = false
+    this.server.on('connection', this.onConnection.bind(this))
+    this.server.on('request', this.wrapRequest.bind(this))
+    return this._listen.apply(this.server, args)
+  }
+
+  close (cb = () => {}) {
+    this.server.removeAllListeners('connection')
+    this.server.removeAllListeners('request')
+    this.server.removeAllListeners('listening')
+    super.close()
+    this._close.call(this.server, cb)
+  }
+
+  destroy (cb) {
+    this.sockets.forEach(socket => {
+      socket.destroy()
+    })
+    super.destroy(cb)
+  }
+}
+
+class BrowserServer extends ServerBase {
+  constructor (client, opts) {
+    super(client, opts)
+
+    this.registration = opts.controller
+    this.workerKeepAliveInterval = null
+    this.workerPortCount = 0
+
+    const scope = new URL(opts.controller.scope)
+    this.pathname = scope.pathname + 'webtorrent'
+    this._address = {
+      port: scope.port,
+      family: 'IPv4', // might be a bad idea?
+      address: scope.hostname
+    }
+
+    this.boundHandler = this.wrapRequest.bind(this)
+    navigator.serviceWorker.addEventListener('message', this.boundHandler)
+    // test if browser supports cancelling sw Readable Streams
+    fetch(`${this.pathname}/cancel/`).then(res => {
+      res.body.cancel()
+    })
+  }
+
+  wrapRequest (event) {
+    const req = event.data
+
+    if (!req?.type === 'webtorrent' || !req.url) return null
+
+    const [port] = event.ports
+    this.onRequest(req, ({ status, headers, body }) => {
+      const asyncIterator = body[Symbol.asyncIterator]?.()
+
+      const cleanup = () => {
+        port.onmessage = null
+        if (body?.destroy) body.destroy()
+        this.workerPortCount--
+        if (!this.workerPortCount) {
+          clearInterval(this.workerKeepAliveInterval)
+          this.workerKeepAliveInterval = null
+        }
+      }
+
+      port.onmessage = async msg => {
+        if (msg.data) {
+          let chunk
+          try {
+            chunk = (await asyncIterator.next()).value
+          } catch (e) {
+            // chunk is yet to be downloaded or it somehow failed, should this be logged?
+          }
+          port.postMessage(chunk)
+          if (!chunk) cleanup()
+          if (!this.workerKeepAliveInterval) {
+            this.workerKeepAliveInterval = setInterval(() => fetch(`${this.pathname}/keepalive/`), keepAliveTime)
+          }
+        } else {
+          cleanup()
+        }
+      }
+      this.workerPortCount++
+      port.postMessage({
+        status,
+        headers,
+        body: asyncIterator ? 'STREAM' : body
+      })
+    })
+  }
+
+  // for compatibility with node version
+  listen (_, cb) {
+    cb()
+  }
+
+  address () {
+    return this._address
+  }
+
+  close (cb) {
+    navigator.serviceWorker.removeEventListener('message', this.boundHandler)
+    super.close(cb)
+  }
+
+  destroy (cb) {
+    super.destroy(cb)
+  }
 }
 
 // NOTE: Arguments must already be HTML-escaped
@@ -274,4 +449,4 @@ function encodeRFC5987 (str) {
     .replace(/%(?:7C|60|5E)/g, unescape)
 }
 
-module.exports = Server
+export { NodeServer, BrowserServer }
