@@ -1,60 +1,61 @@
-/* global Blob */
+import EventEmitter from 'events'
+import fs from 'fs'
+import net from 'net' // browser exclude
+import os from 'os' // browser exclude
+import path from 'path'
+import addrToIPPort from 'addr-to-ip-port'
+import BitField from 'bitfield'
+import CacheChunkStore from 'cache-chunk-store'
+import { chunkStoreWrite } from 'chunk-store-iterator'
+import cpus from 'cpus'
+import debugFactory from 'debug'
+import Discovery from 'torrent-discovery'
+import FSChunkStore from 'fs-chunk-store' // browser: `hybrid-chunk-store`
+import fetch from 'cross-fetch-ponyfill'
+import ImmediateChunkStore from 'immediate-chunk-store'
+import ltDontHave from 'lt_donthave'
+import MemoryChunkStore from 'memory-chunk-store'
+import HybridChunkStore from 'hybrid-chunk-store'
+import joinIterator from 'join-async-iterator'
+import parallel from 'run-parallel'
+import parallelLimit from 'run-parallel-limit'
+import parseTorrent, { toMagnetURI, toTorrentFile, remote } from 'parse-torrent'
+import Piece from 'torrent-piece'
+import queueMicrotask from 'queue-microtask'
+import randomIterate from 'random-iterate'
+import { hash, arr2hex } from 'uint8-util'
+import throughput from 'throughput'
+import utMetadata from 'ut_metadata'
+import utPex from 'ut_pex' // browser exclude
 
-const EventEmitter = require('events')
-const fs = require('fs')
-const net = require('net') // browser exclude
-const os = require('os') // browser exclude
-const path = require('path')
-const addrToIPPort = require('addr-to-ip-port')
-const { default: BitField } = require('bitfield')
-const CacheChunkStore = require('cache-chunk-store')
-const ChunkStoreWriteStream = require('chunk-store-stream/write')
-const cpus = require('cpus')
-const debugFactory = require('debug')
-const Discovery = require('torrent-discovery')
-const FSChunkStore = require('fs-chunk-store') // browser: `memory-chunk-store`
-const get = require('simple-get')
-const ImmediateChunkStore = require('immediate-chunk-store')
-const ltDontHave = require('lt_donthave')
-const MemoryChunkStore = require('memory-chunk-store')
-const MultiStream = require('multistream')
-const parallel = require('run-parallel')
-const parallelLimit = require('run-parallel-limit')
-const parseTorrent = require('parse-torrent')
-const Piece = require('torrent-piece')
-const pump = require('pump')
-const queueMicrotask = require('queue-microtask')
-const randomIterate = require('random-iterate')
-const sha1 = require('simple-sha1')
-const throughput = require('throughput')
-const utMetadata = require('ut_metadata')
-const utPex = require('ut_pex') // browser exclude
+import File from './file.js'
+import Peer from './peer.js'
+import RarityMap from './rarity-map.js'
+import utp from './utp.cjs' // browser exclude
+import WebConn from './webconn.js'
+import { Selections } from './selections.js'
 
-const File = require('./file.js')
-const Peer = require('./peer.js')
-const RarityMap = require('./rarity-map.js')
-const Server = require('./server.js') // browser exclude
-const utp = require('./utp.js') // browser exclude
-const WebConn = require('./webconn.js')
+import VERSION from '../version.cjs'
 
 const debug = debugFactory('webtorrent:torrent')
 const MAX_BLOCK_LENGTH = 128 * 1024
-const PIECE_TIMEOUT = 30000
-const CHOKE_TIMEOUT = 5000
+const PIECE_TIMEOUT = 30_000
+const CHOKE_TIMEOUT = 5_000
 const SPEED_THRESHOLD = 3 * Piece.BLOCK_LENGTH
 
 const PIPELINE_MIN_DURATION = 0.5
 const PIPELINE_MAX_DURATION = 1
 
-const RECHOKE_INTERVAL = 10000 // 10 seconds
+const RECHOKE_INTERVAL = 10_000 // 10 seconds
 const RECHOKE_OPTIMISTIC_DURATION = 2 // 30 seconds
+
+const DEFAULT_NO_PEERS_INTERVAL = 30_000 // 30 seconds
 
 // IndexedDB chunk stores used in the browser benefit from high concurrency
 const FILESYSTEM_CONCURRENCY = process.browser ? cpus().length : 2
 
-const RECONNECT_WAIT = [1000, 5000, 15000]
+const RECONNECT_WAIT = [1_000, 5_000, 15_000]
 
-const VERSION = require('../package.json').version
 const USER_AGENT = `WebTorrent/${VERSION} (https://webtorrent.io)`
 
 let TMP
@@ -64,7 +65,9 @@ try {
   TMP = path.join(typeof os.tmpdir === 'function' ? os.tmpdir() : '/', 'webtorrent')
 }
 
-class Torrent extends EventEmitter {
+const IDLE_CALLBACK = typeof window !== 'undefined' && typeof window.requestIdleCallback === 'function' && window.requestIdleCallback
+
+export default class Torrent extends EventEmitter {
   constructor (torrentId, client, opts) {
     super()
 
@@ -76,11 +79,16 @@ class Torrent extends EventEmitter {
 
     this.path = opts.path || TMP
     this.addUID = opts.addUID || false
+    this.rootDir = opts.rootDir || null
     this.skipVerify = !!opts.skipVerify
     this._store = opts.store || FSChunkStore
     this._preloadedStore = opts.preloadedStore || null
     this._storeCacheSlots = opts.storeCacheSlots !== undefined ? opts.storeCacheSlots : 20
     this._destroyStoreOnDestroy = opts.destroyStoreOnDestroy || false
+    this.store = null
+    this.storeOpts = opts.storeOpts
+    this.alwaysChokeSeeders = opts.alwaysChokeSeeders ?? true
+
     this._getAnnounceOpts = opts.getAnnounceOpts
 
     // if defined, `opts.private` overrides default privacy of torrent
@@ -96,6 +104,9 @@ class Torrent extends EventEmitter {
     this._rechokeOptimisticWire = null
     this._rechokeOptimisticTime = 0
     this._rechokeIntervalId = null
+    this._noPeersIntervalId = null
+    this._noPeersIntervalTime = opts.noPeersIntervalTime ? opts.noPeersIntervalTime * 1000 : DEFAULT_NO_PEERS_INTERVAL
+    this._startAsDeselected = opts.deselect || false
 
     this.ready = false
     this.destroyed = false
@@ -103,13 +114,13 @@ class Torrent extends EventEmitter {
     this.done = false
 
     this.metadata = null
-    this.store = null
-    this.storeOpts = opts.storeOpts
     this.files = []
+
+    // Pieces that need to be downloaded, indexed by piece index
     this.pieces = []
 
     this._amInterested = false
-    this._selections = []
+    this._selections = new Selections()
     this._critical = []
 
     this.wires = [] // open wires (added *after* handshake)
@@ -179,12 +190,9 @@ class Torrent extends EventEmitter {
 
   get numPeers () { return this.wires.length }
 
-  get torrentFileBlobURL () {
-    if (typeof window === 'undefined') throw new Error('browser-only property')
+  get torrentFileBlob () {
     if (!this.torrentFile) return null
-    return URL.createObjectURL(
-      new Blob([this.torrentFile], { type: 'application/x-bittorrent' })
-    )
+    return new Blob([this.torrentFile], { type: 'application/x-bittorrent' })
   }
 
   get _numQueued () {
@@ -199,15 +207,15 @@ class Torrent extends EventEmitter {
     return numConns
   }
 
-  _onTorrentId (torrentId) {
+  async _onTorrentId (torrentId) {
     if (this.destroyed) return
 
     let parsedTorrent
-    try { parsedTorrent = parseTorrent(torrentId) } catch (err) {}
+    try { parsedTorrent = await parseTorrent(torrentId) } catch (err) {}
     if (parsedTorrent) {
       // Attempt to set infoHash property synchronously
       this.infoHash = parsedTorrent.infoHash
-      this._debugId = parsedTorrent.infoHash.toString('hex').substring(0, 7)
+      this._debugId = arr2hex(parsedTorrent.infoHash).substring(0, 7)
       queueMicrotask(() => {
         if (this.destroyed) return
         this._onParsedTorrent(parsedTorrent)
@@ -215,7 +223,7 @@ class Torrent extends EventEmitter {
     } else {
       // If torrentId failed to parse, it could be in a form that requires an async
       // operation, i.e. http/https link, filesystem path, or Blob.
-      parseTorrent.remote(torrentId, (err, parsedTorrent) => {
+      remote(torrentId, (err, parsedTorrent) => {
         if (this.destroyed) return
         if (err) return this._destroy(err)
         this._onParsedTorrent(parsedTorrent)
@@ -256,16 +264,21 @@ class Torrent extends EventEmitter {
   }
 
   _processParsedTorrent (parsedTorrent) {
-    this._debugId = parsedTorrent.infoHash.toString('hex').substring(0, 7)
+    this._debugId = arr2hex(parsedTorrent.infoHash).substring(0, 7)
 
     if (typeof this.private !== 'undefined') {
       // `private` option overrides default, only if it's defined
       parsedTorrent.private = this.private
     }
 
-    if (this.announce) {
+    if (Array.isArray(this.announce)) {
       // Allow specifying trackers via `opts` parameter
       parsedTorrent.announce = parsedTorrent.announce.concat(this.announce)
+    }
+
+    if (this.client.tracker && Array.isArray(this.client.tracker.announce) && !parsedTorrent.private) {
+      // If the client has a default tracker, add it to the announce list if torrent is not private
+      parsedTorrent.announce = parsedTorrent.announce.concat(this.client.tracker.announce)
     }
 
     if (this.client.tracker && global.WEBTORRENT_ANNOUNCE && !parsedTorrent.private) {
@@ -284,8 +297,8 @@ class Torrent extends EventEmitter {
 
     Object.assign(this, parsedTorrent)
 
-    this.magnetURI = parseTorrent.toMagnetURI(parsedTorrent)
-    this.torrentFile = parseTorrent.toTorrentFile(parsedTorrent)
+    this.magnetURI = toMagnetURI(parsedTorrent)
+    this.torrentFile = toTorrentFile(parsedTorrent)
   }
 
   _onListening () {
@@ -329,7 +342,7 @@ class Torrent extends EventEmitter {
 
     // add BEP09 peer-address
     if (this.peerAddresses) {
-      this.peerAddresses.forEach(peer => this.addPeer(peer))
+      this.peerAddresses.forEach(peer => this.addPeer(peer, Peer.SOURCE_MANUAL))
     }
 
     // begin discovering peers via DHT and trackers
@@ -350,24 +363,57 @@ class Torrent extends EventEmitter {
 
     this.discovery.on('peer', (peer, source) => {
       this._debug('peer %s discovered via %s', peer, source)
-      // Don't create new outgoing TCP connections when torrent is done
-      if (typeof peer === 'string' && this.done) return
-      this.addPeer(peer)
+      // Don't create new outgoing connections when torrent is done and seedOutgoingConnections is false.
+      if (!this.client.seedOutgoingConnections && this.done) {
+        this._debug('ignoring peer %s: torrent is done and seedOutgoingConnections is false', peer)
+        return
+      }
+      this.addPeer(peer, source)
     })
 
     this.discovery.on('trackerAnnounce', () => {
       this.emit('trackerAnnounce')
-      if (this.numPeers === 0) this.emit('noPeers', 'tracker')
     })
 
     this.discovery.on('dhtAnnounce', () => {
       this.emit('dhtAnnounce')
-      if (this.numPeers === 0) this.emit('noPeers', 'dht')
     })
 
     this.discovery.on('warning', (err) => {
       this.emit('warning', err)
     })
+
+    this._noPeersIntervalId = setInterval(() => {
+      if (this.destroyed) return
+
+      const counters = {
+        [Peer.SOURCE_TRACKER]: {
+          enabled: !!this.client.tracker,
+          numPeers: 0
+        },
+        [Peer.SOURCE_DHT]: {
+          enabled: !!this.client.dht,
+          numPeers: 0
+        },
+        [Peer.SOURCE_LSD]: {
+          enabled: !!this.client.lsd,
+          numPeers: 0
+        },
+        [Peer.SOURCE_UT_PEX]: {
+          enabled: (this.client.utPex && typeof utPex === 'function'),
+          numPeers: 0
+        }
+      }
+      for (const peer of Object.values(this._peers)) {
+        const counter = counters[peer.source]
+        if (typeof counter !== 'undefined') counter.numPeers++
+      }
+      for (const source of Object.keys(counters)) {
+        const counter = counters[source]
+        if (counter.enabled && counter.numPeers === 0) this.emit('noPeers', source)
+      }
+    }, this._noPeersIntervalTime)
+    if (this._noPeersIntervalId.unref) this._noPeersIntervalId.unref()
   }
 
   _getMetadataFromServer () {
@@ -376,79 +422,79 @@ class Torrent extends EventEmitter {
 
     const urls = Array.isArray(this.xs) ? this.xs : [this.xs]
 
+    self._xsRequestsController = new AbortController()
+
+    const signal = self._xsRequestsController.signal
+
     const tasks = urls.map(url => cb => {
       getMetadataFromURL(url, cb)
     })
     parallel(tasks)
 
-    function getMetadataFromURL (url, cb) {
+    async function getMetadataFromURL (url, cb) {
       if (url.indexOf('http://') !== 0 && url.indexOf('https://') !== 0) {
         self.emit('warning', new Error(`skipping non-http xs param: ${url}`))
         return cb(null)
       }
 
       const opts = {
-        url,
         method: 'GET',
         headers: {
           'user-agent': USER_AGENT
-        }
+        },
+        signal
       }
-      let req
+      let res
       try {
-        req = get.concat(opts, onResponse)
+        res = await fetch(url, opts)
       } catch (err) {
-        self.emit('warning', new Error(`skipping invalid url xs param: ${url}`))
+        self.emit('warning', new Error(`http error from xs param: ${url}`))
         return cb(null)
       }
 
-      self._xsRequests.push(req)
+      if (self.destroyed) return cb(null)
+      if (self.metadata) return cb(null)
 
-      function onResponse (err, res, torrent) {
-        if (self.destroyed) return cb(null)
-        if (self.metadata) return cb(null)
-
-        if (err) {
-          self.emit('warning', new Error(`http error from xs param: ${url}`))
-          return cb(null)
-        }
-        if (res.statusCode !== 200) {
-          self.emit('warning', new Error(`non-200 status code ${res.statusCode} from xs param: ${url}`))
-          return cb(null)
-        }
-
-        let parsedTorrent
-        try {
-          parsedTorrent = parseTorrent(torrent)
-        } catch (err) {}
-
-        if (!parsedTorrent) {
-          self.emit('warning', new Error(`got invalid torrent file from xs param: ${url}`))
-          return cb(null)
-        }
-
-        if (parsedTorrent.infoHash !== self.infoHash) {
-          self.emit('warning', new Error(`got torrent file with incorrect info hash from xs param: ${url}`))
-          return cb(null)
-        }
-
-        self._onMetadata(parsedTorrent)
-        cb(null)
+      if (res.status !== 200) {
+        self.emit('warning', new Error(`non-200 status code ${res.status} from xs param: ${url}`))
+        return cb(null)
       }
+      let torrent
+      try {
+        torrent = new Uint8Array(await res.arrayBuffer())
+      } catch (e) {
+        self.emit('warning', e)
+        return cb(null)
+      }
+
+      let parsedTorrent
+      try {
+        parsedTorrent = await parseTorrent(torrent)
+      } catch (err) {}
+
+      if (!parsedTorrent) {
+        self.emit('warning', new Error(`got invalid torrent file from xs param: ${url}`))
+        return cb(null)
+      }
+
+      if (parsedTorrent.infoHash !== self.infoHash) {
+        self.emit('warning', new Error(`got torrent file with incorrect info hash from xs param: ${url}`))
+        return cb(null)
+      }
+      self._onMetadata(parsedTorrent)
+      cb(null)
     }
   }
 
   /**
    * Called when the full torrent metadata is received.
    */
-  _onMetadata (metadata) {
+  async _onMetadata (metadata) {
     if (this.metadata || this.destroyed) return
     this._debug('got metadata')
 
-    this._xsRequests.forEach(req => {
-      req.abort()
-    })
-    this._xsRequests = []
+    this._xsRequestsController?.abort()
+    this._xsRequestsController = null
 
     let parsedTorrent
     if (metadata && metadata.infoHash) {
@@ -456,7 +502,7 @@ class Torrent extends EventEmitter {
       parsedTorrent = metadata
     } else {
       try {
-        parsedTorrent = parseTorrent(metadata)
+        parsedTorrent = await parseTorrent(metadata)
       } catch (err) {
         return this._destroy(err)
       }
@@ -485,12 +531,14 @@ class Torrent extends EventEmitter {
         files: this.files,
         length: this.length,
         name: this.name + ' - ' + this.infoHash.slice(0, 8),
-        addUID: this.addUID
+        addUID: this.addUID,
+        rootDir: this.rootDir,
+        max: this._storeCacheSlots
       })
     }
 
     // don't use the cache if the store is already in memory
-    if (this._storeCacheSlots > 0 && !(rawStore instanceof MemoryChunkStore)) {
+    if (this._storeCacheSlots > 0 && !(rawStore instanceof MemoryChunkStore || rawStore instanceof HybridChunkStore)) {
       rawStore = new CacheChunkStore(rawStore, {
         max: this._storeCacheSlots
       })
@@ -501,18 +549,16 @@ class Torrent extends EventEmitter {
     )
 
     // Select only specified files (BEP53) http://www.bittorrent.org/beps/bep_0053.html
-    if (this.so) {
+    if (this.so && !this._startAsDeselected) {
       this.files.forEach((v, i) => {
         if (this.so.includes(i)) {
           this.files[i].select()
-        } else {
-          this.files[i].deselect()
         }
       })
     } else {
       // start off selecting the entire torrent with low priority
-      if (this.pieces.length !== 0) {
-        this.select(0, this.pieces.length - 1, false)
+      if (this.pieces.length !== 0 && !this._startAsDeselected) {
+        this.select(0, this.pieces.length - 1, 0)
       }
     }
 
@@ -596,21 +642,22 @@ class Torrent extends EventEmitter {
       if (index === this.pieces.length - 1) {
         getOpts.length = this.lastPieceLength
       }
-      this.store.get(index, getOpts, (err, buf) => {
+      this.store.get(index, getOpts, async (err, buf) => {
         if (this.destroyed) return cb(new Error('torrent is destroyed'))
 
         if (err) return queueMicrotask(() => cb(null)) // ignore error
-        sha1(buf, hash => {
-          if (this.destroyed) return cb(new Error('torrent is destroyed'))
 
-          if (hash === this._hashes[index]) {
-            this._debug('piece verified %s', index)
-            this._markVerified(index)
-          } else {
-            this._debug('piece invalid %s', index)
-          }
-          cb(null)
-        })
+        const hex = await hash(buf, 'hex')
+        if (this.destroyed) return cb(new Error('torrent is destroyed'))
+
+        if (hex === this._hashes[index]) {
+          this._debug('piece verified %s', index)
+          this._markVerified(index)
+        } else {
+          this._markUnverified(index)
+          this._debug('piece invalid %s', index)
+        }
+        cb(null)
       })
     }), FILESYSTEM_CONCURRENCY, cb)
   }
@@ -640,6 +687,19 @@ class Torrent extends EventEmitter {
     this.pieces[index] = null
     this._reservations[index] = null
     this.bitfield.set(index, true)
+    this.emit('verified', index)
+  }
+
+  _markUnverified (index) {
+    const len = (index === this.pieces.length - 1)
+      ? this.lastPieceLength
+      : this.pieceLength
+    this.pieces[index] = new Piece(len)
+    this.bitfield.set(index, false)
+    if (!this._startAsDeselected) this.select(index, index, 1)
+    this.files.forEach(file => {
+      if (file.done && file.includes(index)) file.done = false
+    })
   }
 
   _hasAllPieces () {
@@ -707,11 +767,13 @@ class Torrent extends EventEmitter {
 
     this.client._remove(this)
 
+    this._selections.clear()
+
     clearInterval(this._rechokeIntervalId)
 
-    this._xsRequests.forEach(req => {
-      req.abort()
-    })
+    clearInterval(this._noPeersIntervalId)
+
+    this._xsRequestsController?.abort()
 
     if (this._rarityMap) {
       this._rarityMap.destroy()
@@ -778,7 +840,7 @@ class Torrent extends EventEmitter {
     this._xsRequests = null
   }
 
-  addPeer (peer) {
+  addPeer (peer, source) {
     if (this.destroyed) throw new Error('torrent is destroyed')
     if (!this.infoHash) throw new Error('addPeer() must not be called before the `infoHash` event')
 
@@ -810,7 +872,7 @@ class Torrent extends EventEmitter {
     // if the utp connection fails to connect, then it is replaced with a tcp connection to the same ip:port
 
     const type = (this.client.utp && this._isIPv4(host)) ? 'utp' : 'tcp'
-    const wasAdded = !!this._addPeer(peer, type)
+    const wasAdded = !!this._addPeer(peer, type, source)
 
     if (wasAdded) {
       this.emit('peer', peer)
@@ -820,7 +882,7 @@ class Torrent extends EventEmitter {
     return wasAdded
   }
 
-  _addPeer (peer, type) {
+  _addPeer (peer, type, source) {
     if (this.destroyed) {
       if (typeof peer !== 'string') peer.destroy()
       return null
@@ -950,23 +1012,24 @@ class Torrent extends EventEmitter {
   }
 
   removePeer (peer) {
-    const id = (peer && peer.id) || peer
-    peer = this._peers[id]
+    const id = peer?.id || peer
+    if (peer && !peer.id) peer = this._peers?.[id]
 
     if (!peer) return
+    peer.destroy()
+
+    if (this.destroyed) return
 
     this._debug('removePeer %s', id)
 
     delete this._peers[id]
     this._peersLength -= 1
 
-    peer.destroy()
-
     // If torrent swarm was at capacity before, try to open a new connection now
     this._drain()
   }
 
-  select (start, end, priority, notify) {
+  _select (start, end, priority, notify, isStreamSelection = false) {
     if (this.destroyed) throw new Error('torrent is destroyed')
 
     if (start < 0 || end < start || this.pieces.length <= end) {
@@ -976,12 +1039,13 @@ class Torrent extends EventEmitter {
 
     this._debug('select %s-%s (priority %s)', start, end, priority)
 
-    this._selections.push({
+    this._selections.insert({
       from: start,
       to: end,
       offset: 0,
       priority,
-      notify: notify || noop
+      notify,
+      isStreamSelection
     })
 
     this._selections.sort((a, b) => b.priority - a.priority)
@@ -989,21 +1053,22 @@ class Torrent extends EventEmitter {
     this._updateSelections()
   }
 
-  deselect (start, end, priority) {
+  select (start, end, priority, notify) {
+    this._select(start, end, priority, notify, false)
+  }
+
+  _deselect (from, to, isStreamSelection = false) {
     if (this.destroyed) throw new Error('torrent is destroyed')
 
-    priority = Number(priority) || 0
-    this._debug('deselect %s-%s (priority %s)', start, end, priority)
+    this._debug('deselect %s-%s', from, to)
 
-    for (let i = 0; i < this._selections.length; ++i) {
-      const s = this._selections[i]
-      if (s.from === start && s.to === end && s.priority === priority) {
-        this._selections.splice(i, 1)
-        break
-      }
-    }
+    this._selections.remove({ from, to, isStreamSelection })
 
     this._updateSelections()
+  }
+
+  deselect (start, end) {
+    this._deselect(start, end, false)
   }
 
   critical (start, end) {
@@ -1078,14 +1143,14 @@ class Torrent extends EventEmitter {
     }
 
     // use ut_pex extension if the torrent is not flagged as private
-    if (typeof utPex === 'function' && !this.private) {
+    if (this.client.utPex && typeof utPex === 'function' && !this.private) {
       wire.use(utPex())
 
       wire.ut_pex.on('peer', peer => {
         // Only add potential new peers when we're not seeding
         if (this.done) return
         this._debug('ut_pex: got peer: %s (from %s)', peer, addr)
-        this.addPeer(peer)
+        this.addPeer(peer, Peer.SOURCE_UT_PEX)
       })
 
       wire.ut_pex.on('dropped', peer => {
@@ -1141,7 +1206,7 @@ class Torrent extends EventEmitter {
         if (!wire.peerPieces.get(i)) return
       }
       wire.isSeeder = true
-      wire.choke() // always choke seeders
+      if (this.alwaysChokeSeeders) wire.choke() // always choke seeders
     }
 
     wire.on('bitfield', () => {
@@ -1165,7 +1230,7 @@ class Torrent extends EventEmitter {
     // fast extension (BEP6)
     wire.on('have-all', () => {
       wire.isSeeder = true
-      wire.choke() // always choke seeders
+      if (this.alwaysChokeSeeders) wire.choke() // always choke seeders
       this._update()
       this._updateWireInterest(wire)
     })
@@ -1249,8 +1314,7 @@ class Torrent extends EventEmitter {
    * Garbage collect selections with respect to the store's current state.
    */
   _gcSelections () {
-    for (let i = 0; i < this._selections.length; ++i) {
-      const s = this._selections[i]
+    for (const s of this._selections) {
       const oldOffset = s.offset
 
       // check for newly downloaded pieces in selection
@@ -1258,14 +1322,12 @@ class Torrent extends EventEmitter {
         s.offset += 1
       }
 
-      if (oldOffset !== s.offset) s.notify()
+      if (oldOffset !== s.offset) s.notify?.()
       if (s.to !== s.from + s.offset) continue
       if (!this.bitfield.get(s.from + s.offset)) continue
 
-      this._selections.splice(i, 1) // remove fully downloaded selection
-      i -= 1 // decrement i to offset splice
-
-      s.notify()
+      s.remove() // remove fully downloaded selection
+      s.notify?.()
       this._updateInterest()
     }
 
@@ -1303,23 +1365,20 @@ class Torrent extends EventEmitter {
    * Heartbeat to update all peers and their requests.
    */
   _update () {
-    if (this.destroyed) return
+    if (IDLE_CALLBACK) {
+      IDLE_CALLBACK(() => this._updateWireWrapper(), { timeout: 250 })
+    } else {
+      this._updateWireWrapper()
+    }
+  }
 
+  _updateWireWrapper () {
+    if (this.destroyed) return
     // update wires in random order for better request distribution
     const ite = randomIterate(this.wires)
     let wire
     while ((wire = ite())) {
-      this._updateWireWrapper(wire)
-    }
-  }
-
-  _updateWireWrapper (wire) {
-    const self = this
-
-    if (typeof window !== 'undefined' && typeof window.requestIdleCallback === 'function') {
-      window.requestIdleCallback(() => { self._updateWire(wire) }, { timeout: 250 })
-    } else {
-      self._updateWire(wire)
+      this._updateWire(wire)
     }
   }
 
@@ -1327,6 +1386,7 @@ class Torrent extends EventEmitter {
    * Attempts to update a peer's requests
    */
   _updateWire (wire) {
+    if (wire.destroyed) return false
     // to allow function hoisting
     const self = this
 
@@ -1376,7 +1436,7 @@ class Torrent extends EventEmitter {
 
       let i = self._selections.length
       while (i--) {
-        const next = self._selections[i]
+        const next = self._selections.get(i)
         let piece
         if (self.strategy === 'rarest') {
           const start = next.from + next.offset
@@ -1437,12 +1497,10 @@ class Torrent extends EventEmitter {
 
     function shufflePriority (i) {
       let last = i
-      for (let j = i; j < self._selections.length && self._selections[j].priority; j++) {
+      for (let j = i; j < self._selections.length && self._selections.get(j).priority; j++) {
         last = j
       }
-      const tmp = self._selections[i]
-      self._selections[i] = self._selections[last]
-      self._selections[last] = tmp
+      self._selections.swap(i, last)
     }
 
     function trySelectWire (hotswap) {
@@ -1450,7 +1508,7 @@ class Torrent extends EventEmitter {
       const rank = speedRanker()
 
       for (let i = 0; i < self._selections.length; i++) {
-        const next = self._selections[i]
+        const next = self._selections.get(i)
 
         let piece
         if (self.strategy === 'rarest') {
@@ -1668,7 +1726,7 @@ class Torrent extends EventEmitter {
     const chunkOffset = piece.chunkOffset(reservation)
     const chunkLength = isWebSeed ? piece.chunkLengthRemaining(reservation) : piece.chunkLength(reservation)
 
-    wire.request(index, chunkOffset, chunkLength, function onChunk (err, chunk) {
+    wire.request(index, chunkOffset, chunkLength, async function onChunk (err, chunk) {
       if (self.destroyed) return
 
       // TODO: what is this for?
@@ -1700,34 +1758,33 @@ class Torrent extends EventEmitter {
 
       // TODO: might need to set self.pieces[index] = null here since sha1 is async
 
-      sha1(buf, hash => {
-        if (self.destroyed) return
+      const hex = await hash(buf, 'hex')
+      if (self.destroyed) return
 
-        if (hash === self._hashes[index]) {
-          self._debug('piece verified %s', index)
+      if (hex === self._hashes[index]) {
+        self._debug('piece verified %s', index)
 
-          self.store.put(index, buf, err => {
-            if (err) {
-              self._destroy(err)
-              return
-            } else {
-              self.pieces[index] = null
-              self._markVerified(index)
-              self.wires.forEach(wire => {
-                wire.have(index)
-              })
-            }
-            // We also check `self.destroyed` since `torrent.destroy()` could have been
-            // called in the `torrent.on('done')` handler, triggered by `_checkDone()`.
-            if (self._checkDone() && !self.destroyed) self.discovery.complete()
-            onUpdateTick()
-          })
-        } else {
-          self.pieces[index] = new Piece(piece.length)
-          self.emit('warning', new Error(`Piece ${index} failed verification`))
+        self.store.put(index, buf, err => {
+          if (err) {
+            self._destroy(err)
+            return
+          } else {
+            self.pieces[index] = null
+            self._markVerified(index)
+            self.wires.forEach(wire => {
+              wire.have(index)
+            })
+          }
+          // We also check `self.destroyed` since `torrent.destroy()` could have been
+          // called in the `torrent.on('done')` handler, triggered by `_checkDone()`.
+          if (self._checkDone() && !self.destroyed) self.discovery.complete()
           onUpdateTick()
-        }
-      })
+        })
+      } else {
+        self.pieces[index] = new Piece(piece.length)
+        self.emit('warning', new Error(`Piece ${index} failed verification`))
+        onUpdateTick()
+      }
     })
 
     function onUpdateTick () {
@@ -1777,30 +1834,22 @@ class Torrent extends EventEmitter {
     return done
   }
 
-  load (streams, cb) {
+  async load (streams, cb) {
     if (this.destroyed) throw new Error('torrent is destroyed')
     if (!this.ready) return this.once('ready', () => { this.load(streams, cb) })
 
     if (!Array.isArray(streams)) streams = [streams]
     if (!cb) cb = noop
 
-    const readable = new MultiStream(streams)
-    const writable = new ChunkStoreWriteStream(this.store, this.pieceLength)
-
-    pump(readable, writable, err => {
-      if (err) return cb(err)
+    try {
+      await chunkStoreWrite(this.store, joinIterator(streams), { chunkLength: this.pieceLength })
       this._markAllVerified()
       this._checkDone()
       cb(null)
-    })
-  }
-
-  createServer (requestListener) {
-    if (typeof Server !== 'function') throw new Error('node.js-only method')
-    if (this.destroyed) throw new Error('torrent is destroyed')
-    const server = new Server(this, requestListener)
-    this._servers.push(server)
-    return server
+    } catch (err) {
+      cb(err)
+      return err
+    }
   }
 
   pause () {
@@ -1847,7 +1896,7 @@ class Torrent extends EventEmitter {
       port: parts[1]
     }
 
-    if (this.client.utp && peer.type === 'utpOutgoing') {
+    if (this.client.utp && peer.type === Peer.TYPE_UTP_OUTGOING) {
       peer.conn = utp.connect(opts.port, opts.host)
     } else {
       peer.conn = net.connect(opts)
@@ -1855,7 +1904,7 @@ class Torrent extends EventEmitter {
 
     const conn = peer.conn
 
-    conn.once('connect', () => { peer.onConnect() })
+    conn.once('connect', () => { if (!this.destroyed) peer.onConnect() })
     conn.once('error', err => { peer.destroy(err) })
     peer.startConnectTimeout()
 
@@ -1865,7 +1914,7 @@ class Torrent extends EventEmitter {
 
       if (peer.retries >= RECONNECT_WAIT.length) {
         if (this.client.utp) {
-          const newPeer = this._addPeer(peer.addr, 'tcp')
+          const newPeer = this._addPeer(peer.addr, 'tcp', peer.source)
           if (newPeer) newPeer.retries = 0
         } else {
           this._debug(
@@ -1886,7 +1935,7 @@ class Torrent extends EventEmitter {
         if (this.destroyed) return
         const host = addrToIPPort(peer.addr)[0]
         const type = (this.client.utp && this._isIPv4(host)) ? 'utp' : 'tcp'
-        const newPeer = this._addPeer(peer.addr, type)
+        const newPeer = this._addPeer(peer.addr, type, peer.source)
         if (newPeer) newPeer.retries = peer.retries + 1
       }, ms)
       if (reconnectTimeout.unref) reconnectTimeout.unref()
@@ -1948,5 +1997,3 @@ function randomInt (high) {
 }
 
 function noop () {}
-
-module.exports = Torrent
