@@ -1,12 +1,14 @@
-const { default: BitField } = require('bitfield')
-const debugFactory = require('debug')
-const get = require('simple-get')
-const ltDontHave = require('lt_donthave')
-const sha1 = require('simple-sha1')
-const Wire = require('bittorrent-protocol')
+import BitField from 'bitfield'
+import debugFactory from 'debug'
+import fetch from 'cross-fetch-ponyfill'
+import ltDontHave from 'lt_donthave'
+import { hash, concat } from 'uint8-util'
+import Wire from 'bittorrent-protocol'
+import once from 'once'
+
+import VERSION from '../version.cjs'
 
 const debug = debugFactory('webtorrent:webconn')
-const VERSION = require('../package.json').version
 
 const SOCKET_TIMEOUT = 60000
 const RETRY_DELAY = 10000
@@ -16,26 +18,27 @@ const RETRY_DELAY = 10000
  * @param {string} url web seed url
  * @param {Object} torrent
  */
-class WebConn extends Wire {
+export default class WebConn extends Wire {
   constructor (url, torrent) {
     super()
 
     this.url = url
     this.connId = url // Unique id to deduplicate web seeds
-    this.webPeerId = sha1.sync(url) // Used as the peerId for this fake remote peer
     this._torrent = torrent
 
-    this._init()
+    this._init(url)
   }
 
-  _init () {
+  _init (url) {
     this.setKeepAlive(true)
 
     this.use(ltDontHave())
 
-    this.once('handshake', (infoHash, peerId) => {
+    this.once('handshake', async (infoHash, peerId) => {
+      const hex = await hash(url, 'hex') // Used as the peerId for this fake remote peer
       if (this.destroyed) return
-      this.handshake(infoHash, this.webPeerId)
+      this.handshake(infoHash, hex)
+
       const numPieces = this._torrent.pieces.length
       const bitfield = new BitField(numPieces)
       for (let i = 0; i <= numPieces; i++) {
@@ -76,7 +79,8 @@ class WebConn extends Wire {
     })
   }
 
-  httpRequest (pieceIndex, offset, length, cb) {
+  async httpRequest (pieceIndex, offset, length, cb) {
+    cb = once(cb)
     const pieceOffset = pieceIndex * this._torrent.pieceLength
     const rangeStart = pieceOffset + offset /* offset within whole torrent */
     const rangeEnd = rangeStart + length - 1
@@ -111,101 +115,35 @@ class WebConn extends Wire {
         }
       })
     }
+    let chunks
+    try {
+      chunks = await Promise.all(requests.map(async ({ start, end, url }) => {
+        debug(
+          'Requesting url=%s pieceIndex=%d offset=%d length=%d start=%d end=%d',
+          url, pieceIndex, offset, length, start, end
+        )
+        const res = await fetch(url, {
+          cache: 'no-store',
+          method: 'GET',
+          headers: {
+            'Cache-Control': 'no-store',
+            'user-agent': `WebTorrent/${VERSION} (https://webtorrent.io)`,
+            range: `bytes=${start}-${end}`
+          },
+          signal: AbortSignal.timeout(SOCKET_TIMEOUT)
+        })
+        if (!res.ok) throw new Error(`Unexpected HTTP status code ${res.status}`)
+        const data = new Uint8Array(await res.arrayBuffer())
 
-    // Now make all the HTTP requests we need in order to load this piece
-    // Usually that's one requests, but sometimes it will be multiple
-    // Send requests in parallel and wait for them all to come back
-    let numRequestsSucceeded = 0
-    let hasError = false
-
-    let ret
-    if (requests.length > 1) {
-      ret = Buffer.alloc(length)
-    }
-
-    requests.forEach(request => {
-      const url = request.url
-      const start = request.start
-      const end = request.end
-      debug(
-        'Requesting url=%s pieceIndex=%d offset=%d length=%d start=%d end=%d',
-        url, pieceIndex, offset, length, start, end
-      )
-      const opts = {
-        url,
-        method: 'GET',
-        headers: {
-          'user-agent': `WebTorrent/${VERSION} (https://webtorrent.io)`,
-          range: `bytes=${start}-${end}`
-        },
-        timeout: SOCKET_TIMEOUT
-      }
-      function onResponse (res, data) {
-        if (res.statusCode < 200 || res.statusCode >= 300) {
-          if (hasError) return
-          hasError = true
-          return cb(new Error(`Unexpected HTTP status code ${res.statusCode}`))
-        }
         debug('Got data of length %d', data.length)
 
-        if (requests.length === 1) {
-          // Common case: fetch piece in a single HTTP request, return directly
-          cb(null, data)
-        } else {
-          // Rare case: reconstruct multiple HTTP requests across 2+ files into one
-          // piece buffer
-          data.copy(ret, request.fileOffsetInRange)
-          if (++numRequestsSucceeded === requests.length) {
-            cb(null, ret)
-          }
-        }
-      }
-      get.concat(opts, (err, res, data) => {
-        if (hasError) return
-        if (err) {
-          // Browsers allow HTTP redirects for simple cross-origin
-          // requests but not for requests that require preflight.
-          // Use a simple request to unravel any redirects and get the
-          // final URL.  Retry the original request with the new URL if
-          // it's different.
-          //
-          // This test is imperfect but it's simple and good for common
-          // cases.  It catches all cross-origin cases but matches a few
-          // same-origin cases too.
-          if (typeof window === 'undefined' || url.startsWith(`${window.location.origin}/`)) {
-            hasError = true
-            return cb(err)
-          }
+        return data
+      }))
+    } catch (e) {
+      return cb(e)
+    }
 
-          return get.head(url, (errHead, res) => {
-            if (hasError) return
-            if (errHead) {
-              hasError = true
-              return cb(errHead)
-            }
-            if (res.statusCode < 200 || res.statusCode >= 300) {
-              hasError = true
-              return cb(new Error(`Unexpected HTTP status code ${res.statusCode}`))
-            }
-            if (res.url === url) {
-              hasError = true
-              return cb(err)
-            }
-
-            opts.url = res.url
-            get.concat(opts, (err, res, data) => {
-              if (hasError) return
-              if (err) {
-                hasError = true
-                return cb(err)
-              }
-              onResponse(res, data)
-            })
-          })
-        }
-        onResponse(res, data)
-      })
-    })
+    cb(null, concat(chunks))
   }
 
   destroy () {
@@ -213,5 +151,3 @@ class WebConn extends Wire {
     this._torrent = null
   }
 }
-
-module.exports = WebConn
