@@ -1,6 +1,4 @@
-/* eslint-env serviceworker */
-
-const fileResponse = require('./worker-server.js')
+import fileResponse from './worker-server.js'
 
 self.addEventListener('install', () => {
   self.skipWaiting()
@@ -11,6 +9,6 @@ self.addEventListener('fetch', event => {
   if (res) event.respondWith(res)
 })
 
-self.addEventListener('activate', evt => {
-  evt.waitUntil(self.clients.claim())
+self.addEventListener('activate', () => {
+  self.clients.claim()
 })
