@@ -1,11 +1,11 @@
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
-const WebTorrent = require('../index.js')
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
+import WebTorrent from '../index.js'
 
 test('client.add: magnet uri, utf-8 string', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -13,11 +13,11 @@ test('client.add: magnet uri, utf-8 string', t => {
   const torrent = client.add(fixtures.leaves.magnetURI)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(fixtures.leaves.magnetURI, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.magnetURI, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -27,7 +27,7 @@ test('client.add: magnet uri, utf-8 string', t => {
 test('client.add: torrent file, buffer', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -35,11 +35,12 @@ test('client.add: torrent file, buffer', t => {
   const torrent = client.add(fixtures.leaves.torrent)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(fixtures.leaves.torrent, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.torrent, err => { t.error(err, 'torrent destroyed') })
+    console.log(client.torrents.length)
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -49,7 +50,7 @@ test('client.add: torrent file, buffer', t => {
 test('client.add: info hash, hex string', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -57,11 +58,11 @@ test('client.add: info hash, hex string', t => {
   const torrent = client.add(fixtures.leaves.parsedTorrent.infoHash)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, `magnet:?xt=urn:btih:${fixtures.leaves.parsedTorrent.infoHash}`)
 
-    client.remove(fixtures.leaves.parsedTorrent.infoHash, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.parsedTorrent.infoHash, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -71,7 +72,7 @@ test('client.add: info hash, hex string', t => {
 test('client.add: info hash, buffer', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -79,11 +80,11 @@ test('client.add: info hash, buffer', t => {
   const torrent = client.add(fixtures.leaves.parsedTorrent.infoHashBuffer)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.ok(torrent.magnetURI.indexOf(`magnet:?xt=urn:btih:${fixtures.leaves.parsedTorrent.infoHash}`) === 0)
 
-    client.remove(Buffer.from(fixtures.leaves.parsedTorrent.infoHash, 'hex'), err => { t.error(err, 'torrent destroyed') })
+    await client.remove(Buffer.from(fixtures.leaves.parsedTorrent.infoHash, 'hex'), err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -93,7 +94,7 @@ test('client.add: info hash, buffer', t => {
 test('client.add: parsed torrent, from `parse-torrent`', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -101,11 +102,11 @@ test('client.add: parsed torrent, from `parse-torrent`', t => {
   const torrent = client.add(fixtures.leaves.parsedTorrent)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -115,7 +116,7 @@ test('client.add: parsed torrent, from `parse-torrent`', t => {
 test('client.add: parsed torrent, with string type announce property', t => {
   t.plan(7)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -126,7 +127,7 @@ test('client.add: parsed torrent, with string type announce property', t => {
   const torrent = client.add(parsedTorrent)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
 
     const expectedMagnetURI = `${fixtures.leaves.magnetURI}&tr=${encodeURIComponent('http://tracker.local:80')}`
@@ -135,7 +136,7 @@ test('client.add: parsed torrent, with string type announce property', t => {
     // `torrent.announce` must always be an array
     t.deepEqual(torrent.announce, ['http://tracker.local:80'])
 
-    client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -145,7 +146,7 @@ test('client.add: parsed torrent, with string type announce property', t => {
 test('client.add: parsed torrent, with array type announce property', t => {
   t.plan(7)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -156,7 +157,7 @@ test('client.add: parsed torrent, with array type announce property', t => {
   const torrent = client.add(parsedTorrent)
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
 
     const expectedMagnetURI = `${fixtures.leaves.magnetURI}&tr=${encodeURIComponent('http://tracker.local:80')}&tr=${encodeURIComponent('http://tracker.local:81')}`
@@ -164,7 +165,7 @@ test('client.add: parsed torrent, with array type announce property', t => {
 
     t.deepEqual(torrent.announce, ['http://tracker.local:80', 'http://tracker.local:81'])
 
-    client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.parsedTorrent, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -174,7 +175,7 @@ test('client.add: parsed torrent, with array type announce property', t => {
 test('client.add: invalid torrent id: empty string', t => {
   t.plan(3)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => {
     t.ok(err instanceof Error)
@@ -190,7 +191,7 @@ test('client.add: invalid torrent id: empty string', t => {
 test('client.add: invalid torrent id: short buffer', t => {
   t.plan(3)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => {
     t.ok(err instanceof Error)
@@ -206,7 +207,7 @@ test('client.add: invalid torrent id: short buffer', t => {
 test('client.add: paused torrent', t => {
   t.plan(5)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', (err) => { t.fail(err) })
   client.on('warning', (err) => { t.fail(err) })
@@ -214,10 +215,10 @@ test('client.add: paused torrent', t => {
   const torrent = client.add(fixtures.leaves.magnetURI, { paused: true })
   t.equal(client.torrents.length, 1)
 
-  torrent.on('infoHash', () => {
+  torrent.on('infoHash', async () => {
     t.equal(torrent.paused, true)
 
-    client.remove(fixtures.leaves.magnetURI, err => { t.error(err, 'torrent destroyed') })
+    await client.remove(fixtures.leaves.magnetURI, err => { t.error(err, 'torrent destroyed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
