@@ -1,13 +1,11 @@
-/* global Blob */
-
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
-const WebTorrent = require('../index.js')
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
+import WebTorrent from '../index.js'
 
 test('client.seed: torrent file (Buffer)', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -15,12 +13,12 @@ test('client.seed: torrent file (Buffer)', t => {
   client.seed(fixtures.leaves.content, {
     name: 'Leaves of Grass by Walt Whitman.epub',
     announce: []
-  }, torrent => {
+  }, async torrent => {
     t.equal(client.torrents.length, 1)
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(torrent, err => { t.error(err, 'torrent removed') })
+    await client.remove(torrent, err => { t.error(err, 'torrent removed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -30,7 +28,7 @@ test('client.seed: torrent file (Buffer)', t => {
 test('client.seed: torrent file (Buffer), set name on buffer', t => {
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -38,12 +36,12 @@ test('client.seed: torrent file (Buffer), set name on buffer', t => {
   const buf = Buffer.from(fixtures.leaves.content)
   buf.name = 'Leaves of Grass by Walt Whitman.epub'
 
-  client.seed(buf, { announce: [] }, torrent => {
+  client.seed(buf, { announce: [] }, async torrent => {
     t.equal(client.torrents.length, 1)
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(torrent, err => { t.error(err, 'torrent removed') })
+    await client.remove(torrent, err => { t.error(err, 'torrent removed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
@@ -55,7 +53,7 @@ test('client.seed: torrent file (Blob)', t => {
 
   t.plan(6)
 
-  const client = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client.on('error', err => { t.fail(err) })
   client.on('warning', err => { t.fail(err) })
@@ -63,19 +61,19 @@ test('client.seed: torrent file (Blob)', t => {
   client.seed(new Blob([fixtures.leaves.content]), {
     name: 'Leaves of Grass by Walt Whitman.epub',
     announce: []
-  }, torrent => {
+  }, async torrent => {
     t.equal(client.torrents.length, 1)
     t.equal(torrent.infoHash, fixtures.leaves.parsedTorrent.infoHash)
     t.equal(torrent.magnetURI, fixtures.leaves.magnetURI)
 
-    client.remove(torrent, err => { t.error(err, 'torrent removed') })
+    await client.remove(torrent, err => { t.error(err, 'torrent removed') })
     t.equal(client.torrents.length, 0)
 
     client.destroy(err => { t.error(err, 'client destroyed') })
   })
 })
 
-test('client.seed: duplicate seed)', t => {
+test('client.seed: duplicate seed', t => {
   t.plan(4)
 
   const client = new WebTorrent()
