@@ -1,10 +1,10 @@
-const fs = require('fs')
-const fixtures = require('webtorrent-fixtures')
-const MemoryChunkStore = require('memory-chunk-store')
-const series = require('run-series')
-const test = require('tape')
-const TrackerServer = require('bittorrent-tracker/server')
-const WebTorrent = require('../../index.js')
+import fs from 'fs'
+import fixtures from 'webtorrent-fixtures'
+import MemoryChunkStore from 'memory-chunk-store'
+import series from 'run-series'
+import test from 'tape'
+import { Server as TrackerServer } from 'bittorrent-tracker'
+import WebTorrent from '../../index.js'
 
 test('Download using UDP tracker (via .torrent file)', t => {
   torrentDownloadTest(t, 'udp')
@@ -14,15 +14,23 @@ test('Download using HTTP tracker (via .torrent file)', t => {
   torrentDownloadTest(t, 'http')
 })
 
+test('Download using WS tracker (via .torrent file)', t => {
+  torrentDownloadTest(t, 'ws')
+})
+
+const TRACKER_CONFIG_MAP = {
+  udp: { http: false, ws: false },
+  http: { udp: false, ws: false },
+  ws: { udp: false, http: false, ws: true }
+}
+
 function torrentDownloadTest (t, serverType) {
   t.plan(9)
 
   let trackerStartCount = 0
   const parsedTorrent = Object.assign({}, fixtures.leaves.parsedTorrent)
 
-  const tracker = new TrackerServer(
-    serverType === 'udp' ? { http: false, ws: false } : { udp: false, ws: false }
-  )
+  const tracker = new TrackerServer(TRACKER_CONFIG_MAP[serverType])
 
   tracker.on('error', err => { t.fail(err) })
   tracker.on('warning', err => { t.fail(err) })
@@ -45,10 +53,7 @@ function torrentDownloadTest (t, serverType) {
 
       const port = tracker[serverType].address().port
 
-      const announceUrl = serverType === 'http'
-        ? `http://127.0.0.1:${port}/announce`
-        : `udp://127.0.0.1:${port}`
-
+      const announceUrl = `${serverType}://127.0.0.1:${port}/announce`
       // Overwrite announce with our local tracker
       parsedTorrent.announce = [announceUrl]
 
@@ -75,18 +80,12 @@ function torrentDownloadTest (t, serverType) {
 
       client2.add(parsedTorrent, { store: MemoryChunkStore })
 
-      client2.on('torrent', torrent => {
+      client2.on('torrent', async torrent => {
         let gotBuffer = false
         let torrentDone = false
-
-        torrent.files.forEach(file => {
-          file.getBuffer((err, buf) => {
-            if (err) throw err
-            t.deepEqual(buf, fixtures.leaves.content, 'downloaded correct content')
-            gotBuffer = true
-            maybeDone()
-          })
-        })
+        function maybeDone () {
+          if (gotBuffer && torrentDone) cb(null)
+        }
 
         torrent.once('done', () => {
           t.pass('client2 downloaded torrent from client1')
@@ -94,8 +93,16 @@ function torrentDownloadTest (t, serverType) {
           maybeDone()
         })
 
-        function maybeDone () {
-          if (gotBuffer && torrentDone) cb(null)
+        for (const file of torrent.files) {
+          try {
+            const ab = await file.arrayBuffer()
+            t.deepEqual(new Uint8Array(ab), new Uint8Array(fixtures.leaves.content), 'downloaded correct content')
+          } catch (err) {
+            t.error(err)
+          }
+
+          gotBuffer = true
+          maybeDone()
         }
       })
     }
