@@ -1,7 +1,7 @@
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
-const MemoryChunkStore = require('memory-chunk-store')
-const WebTorrent = require('../../index.js')
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
+import MemoryChunkStore from 'memory-chunk-store'
+import WebTorrent from '../../index.js'
 
 const DOWNLOAD_SPEED_LIMIT = 200 * 1000 // 200 KB/s
 const UPLOAD_SPEED_LIMIT = 200 * 1000 // 200 KB/s
@@ -33,8 +33,10 @@ function testSpeed (t, downloaderOpts, uploaderOpts, cb) {
     // Start downloading
     const torrent = client1.add(fixtures.leaves.parsedTorrent.infoHash, { store: MemoryChunkStore })
 
-    // Manually connect peers
-    torrent.addPeer(`127.0.0.1:${client2.address().port}`)
+    torrent.once('infoHash', () => {
+      // Manually connect peers
+      torrent.addPeer(`127.0.0.1:${client2.address().port}`)
+    })
 
     torrent.on('download', () => {
       downloadSpeeds.push(torrent.downloadSpeed)
