@@ -1,12 +1,12 @@
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
-const WebTorrent = require('../../index.js')
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
+import WebTorrent from '../../index.js'
 
 test('ut_metadata transfer', t => {
   t.plan(6)
 
-  const client1 = new WebTorrent({ dht: false, tracker: false, lsd: false })
-  const client2 = new WebTorrent({ dht: false, tracker: false, lsd: false })
+  const client1 = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
+  const client2 = new WebTorrent({ dht: false, tracker: false, lsd: false, natUpnp: false, natPmp: false })
 
   client1.on('error', err => { t.fail(err) })
   client1.on('warning', err => { t.fail(err) })
@@ -23,6 +23,9 @@ test('ut_metadata transfer', t => {
   client1.add(fixtures.leaves.torrent)
 
   client1.on('torrent', torrent1 => {
+    // TODO: remove these 2 lines once we replace airtap
+    fixtures.leaves.parsedTorrent.info.name = new Uint8Array(fixtures.leaves.parsedTorrent.info.name)
+    fixtures.leaves.parsedTorrent.info.pieces = new Uint8Array(fixtures.leaves.parsedTorrent.info.pieces)
     t.deepEqual(torrent1.info, fixtures.leaves.parsedTorrent.info)
 
     // client2 starts with infohash
