@@ -1,9 +1,8 @@
-
 /**
  * Mapping of torrent pieces to their respective availability in the torrent swarm. Used
  * by the torrent manager for implementing the rarest piece first selection strategy.
  */
-class RarityMap {
+export default class RarityMap {
   constructor (torrent) {
     this._torrent = torrent
     this._numPieces = torrent.pieces.length
@@ -104,5 +103,3 @@ class RarityMap {
     wire._onClose = null
   }
 }
-
-module.exports = RarityMap
